Move tests back to integration.rs
